GFM-159 - updated to use hof-bootstrap
* removed unnec deps
* removed setup from app.js
* installed hof-bootstrap
* added bootstrap setup to app.js
* removed apps/gro/index.js setup
* moved steps config to apps/gro/index.js
