+++ package.json
@@ -27,35 +27,16 @@
     "cover": "istanbul cover _mocha && istanbul check-coverage",
     "sass": "node-sass ./assets/scss/app.scss ./public/css/app.css --include-path ./node_modules",
     "create:public": "mkdir -p ./public/js ./public/css ./public/images",
-    "hof-transpile": "hof-transpiler ./apps/**/translations/src -w --shared ./node_modules/hof-template-partials/translations/src",
+    "hof-transpile": "hof-transpiler ./apps/**/translations/src -w --shared ./node_modules/hof-template-partials/translations",
     "prepareapp": "npm run create:public; npm run sass; npm run browserify; npm run copy:images; npm run hof-transpile",
     "postinstall": "bash -c 'if [[ ${NODE_ENV} != production ]]; then npm run prepareapp; fi;'"
   },
   "author": "HomeOffice",
   "dependencies": {
-    "body-parser": "^1.13.0",
     "browserify": "^12.0.1",
-    "churchill": "0.0.5",
-    "connect-redis-crypto": "^2.1.0",
-    "cookie-parser": "^1.3.5",
-    "express": "^4.12.4",
-    "express-partial-templates": "^0.2.0",
-    "express-session": "^1.11.3",
-    "hof": "UKHomeOffice/hof#so-master",
-    "hof-template-partials": "^1.0.1",
+    "hof-bootstrap": "UKHomeOffice/hof-bootstrap#development",
     "hof-transpiler": "^0.1.0",
-    "hogan-express-strict": "^0.5.4",
-    "hogan.js": "^3.0.2",
-    "i18n-lookup": "^0.1.0",
-    "jquery": "^2.1.4",
-    "lodash": "3.0.1",
-    "node-sass": "^3.7.0",
-    "nodemailer": "^1.4.0",
-    "nodemailer-smtp-transport": "^1.0.3",
-    "nodemailer-stub-transport": "^1.0.0",
-    "redis": "^0.12.1",
-    "typeahead-aria": "^1.0.1",
-    "winston": "^1.0.1"
+    "node-sass": "^3.10.0"
   },
   "devDependencies": {
     "chai": "^3.0.0",
